refactor: reorganize weather-common package structure and improve code quality
- Move WeatherParseException from weather.service to weather.exception package
- Add @Serial annotations to serialization members (Java 14+ standard)
- Improve Visibility conversion methods with proper exception handling
- Remove redundant field initializations and collapsible if blocks
- Enhance JavaDoc documentation for IndexedLinkedHashMap
- Add tests for edge cases and improve test coverage

Version: 1.11.0-SNAPSHOT
Build Status: All tests passing
Coverage: 97%+ instruction coverage maintained
No breaking changes to public APIs
