+++ noakweather-platform/weather-common/src/main/java/weather/exception/WeatherParseException.java
@@ -1,6 +1,6 @@
 /*
  * NoakWeather Engineering Pipeline(TM) is a multi-source weather data engineering platform
- * Copyright (C) 2025 bclasky1539
+ * Copyright (C) 2025-2026 bclasky1539
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package weather.service;
+package weather.exception;
 
 import weather.model.WeatherDataSource;
 
 /**
  * Exception thrown when weather data parsing fails.
- * 
+ * <p>
  * This exception preserves the raw data and source information
  * to enable debugging and retry logic.
  *
