+++ noakweather-platform/weather-common/src/main/java/weather/utils/IndexedLinkedHashMap.java
@@ -1,6 +1,6 @@
 /*
  * NoakWeather Engineering Pipeline(TM) is a multi-source weather data engineering platform
- * Copyright (C) 2025 bclasky1539
+ * Copyright (C) 2025-2026 bclasky1539
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@ package weather.utils;
 
 import java.io.IOException;
 import java.io.ObjectInputStream;
+import java.io.Serial;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.Objects;
@@ -36,6 +37,7 @@ public class IndexedLinkedHashMap<K, V> extends LinkedHashMap<K, V> {
     /**
      *
      */
+    @Serial
     private static final long serialVersionUID = 1L;
 
     private transient ArrayList<K> indexList = new ArrayList<>();
@@ -48,10 +50,12 @@ public class IndexedLinkedHashMap<K, V> extends LinkedHashMap<K, V> {
     }
 
     /**
-     * Equals method override.
+     * Compares the specified object with this map for equality.
+     * Returns true if the given object is also an IndexedLinkedHashMap with the same
+     * mappings and insertion order.
      *
-     * @param obj
-     * @return value
+     * @param obj the object to be compared for equality with this map
+     * @return true if the specified object is equal to this map
      */
     @Override
     @SuppressWarnings("AccessingNonPublicFieldOfAnotherObject")
@@ -71,9 +75,9 @@ public class IndexedLinkedHashMap<K, V> extends LinkedHashMap<K, V> {
     }
 
     /**
-     * Sets the hash.
+     * Returns the hash code value for this map.
      *
-     * @return value
+     * @return the hash code value for this map
      */
     @Override
     public int hashCode() {
@@ -81,11 +85,12 @@ public class IndexedLinkedHashMap<K, V> extends LinkedHashMap<K, V> {
     }
 
     /**
-     * Sets the locale of the bundle.
+     * Associates the specified value with the specified key in this map.
+     * If the key is new, it is added to the index list to maintain insertion order.
      *
-     * @param key
-     * @param val
-     * @return value
+     * @param key the key with which the specified value is to be associated
+     * @param val the value to be associated with the specified key
+     * @return the previous value associated with key, or null if there was no mapping for key
      */
     @Override
     public V put(K key, V val) {
@@ -96,11 +101,11 @@ public class IndexedLinkedHashMap<K, V> extends LinkedHashMap<K, V> {
     }
 
     /**
-     * Remove the specified key and update the index list.
-     * Only remove from indexList if key was actually in the map.
+     * Removes the mapping for the specified key from this map if present.
+     * Also removes the key from the index list to maintain consistency.
      *
-     * @param key
-     * @return the previous value associated with key, or null
+     * @param key the key whose mapping is to be removed from the map
+     * @return the previous value associated with key, or null if there was no mapping for key
      */
     @Override
     public V remove(Object key) {
@@ -110,9 +115,9 @@ public class IndexedLinkedHashMap<K, V> extends LinkedHashMap<K, V> {
         }
         return value;
     }
-    
+
     /**
-     * Clear all entries and reset the index list.
+     * Removes all mappings from this map and clears the index list.
      */
     @Override
     public void clear() {
@@ -121,47 +126,49 @@ public class IndexedLinkedHashMap<K, V> extends LinkedHashMap<K, V> {
     }
 
     /**
-     * Get the value at the specified index position.
+     * Returns the value at the specified index position in insertion order.
      *
-     * @param i
-     * @return index
+     * @param i the index of the value to return (0-based)
+     * @return the value at the specified index position
+     * @throws IndexOutOfBoundsException if the index is out of range
      */
     public V getValueAtIndex(int i) {
         return super.get(indexList.get(i));
     }
 
     /**
-     * Get the key at the specified index position.
+     * Returns the key at the specified index position in insertion order.
      *
-     * @param i
-     * @return index
+     * @param i the index of the key to return (0-based)
+     * @return the key at the specified index position
+     * @throws IndexOutOfBoundsException if the index is out of range
      */
     public K getKeyAtIndex(int i) {
         return indexList.get(i);
     }
 
     /**
-     * Get the value at index of key.
+     * Returns the index position of the specified key in insertion order.
      *
-     * @param key
-     * @return index of key
+     * @param key the key whose index is to be returned
+     * @return the index of the key, or -1 if the key is not present
      */
     public int getIndexOf(K key) {
         return indexList.indexOf(key);
     }
 
     /**
-     * Called automatically during de-serialization - no impact on calling code.
+     * Reconstitutes the IndexedLinkedHashMap instance from a stream (deserializes it).
+     * Rebuilds the index list from the deserialized map contents to maintain insertion order.
      *
-     * @param in
-     * @return
+     * @param in the ObjectInputStream from which to read the object
+     * @throws IOException if an I/O error occurs
+     * @throws ClassNotFoundException if the class of a serialized object cannot be found
      */
+    @Serial
     private void readObject(ObjectInputStream in) throws IOException, ClassNotFoundException {
         in.defaultReadObject();
         // Rebuild the index from the deserialized map contents
-        indexList = new ArrayList<>();
-        for (K key : this.keySet()) {
-            indexList.add(key);
-        }
+        indexList = new ArrayList<>(this.keySet());
     }
 }
